day6: Chronal Coordinates
Nothing terribly clever, just brute force. Iterating through the grid
squares rather than doing distances for each point is a pretty huge
speedup (~4x) which is kind of a surprise. I didn't have any time to dig
into it, but I'd be interested in knowing why.

Of course, not building the grid at all is the fastest approach for
part2 I'm sure.