day9: marble mania
initial solution is solve. It worked but was a bit annoying to write and
get all the arithmetic working properly for. It finished part1 in what
I thought was a reasonable amount of time, but didn't seem likely to
finish part2. multiplying the number by 10 instead of 100 took 30s
(instead of ~250ms for part1), and I had a hunch it was mostly due
to shuffling all the vec elements to and fro.

solve_better is more involved and required writing the new circular
buffer data structure, but I think the solution itself is much simpler.
It finishes part2 in less time than solve takes for part1.