day11: Chronal Charge
Did the first part and got sort of distracted and didn't have much time
to do the second so just wrote the easy brute force one. Surprisingly,
it only took about a minute to run, and it was down to ~15 seconds using
i32 instead of i64s and i8s.

The other two I did a bit later on. The transposed one takes about half
the time of the other (though obviously for twice the memory).