day23: my triumphant return
Been quite busy but thought today would be quite quick. The first part
is trivial, but the second part is a bit of a mess. I'd done some z3 in
haskell for some different puzzles, but it's not nearly as nice in rust
tbh. It's not really very well abstracted, and it takes about 2 minutes
to run on my laptop. But it does get through it eventually...